Forward declare serialiser types
